refactor(app): move InlineCalibrationWarning up a level
It will be used in the pipettes card soon, so this makes it so we don't
need cross-tree imports.
